Remove library methods deprecated in M3
* Remove `scala.deriving.{ArrayProduct, EmptyProduct, productElement}`
* Remove `scala.quoted.Quotes.reflect.{Tree.of, Term.of}`
* Remove `scala.quoted.{Unlifted, Const, Consts}`
