+++ tests/run-macros/tasty-extractors-constants-1.check
@@ -2,4 +2,3 @@
 4
 abc
 null
-OK
